移植成功ROS2格式

+++ src/Fast-Planner/fast_planner/poly_traj/include/poly_traj/polynomial_traj.h
@@ -63,7 +63,7 @@ public:
   void init() {
     num_seg = times.size();
     time_sum = 0.0;
-    for (int i = 0; i < times.size(); ++i) {
+    for (size_t i = 0; i < times.size(); ++i) {
       time_sum += times[i];
     }
   }
@@ -79,7 +79,7 @@ public:
     /* evaluation */
     int order = cxs[idx].size();
     Eigen::VectorXd cx(order), cy(order), cz(order), tv(order);
-    for (int i = 0; i < order; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(order); ++i) {
       cx(i) = cxs[idx][i], cy(i) = cys[idx][i], cz(i) = czs[idx][i];
       tv(order - 1 - i) = std::pow(t, double(i));
     }
@@ -102,14 +102,14 @@ public:
     Eigen::VectorXd vx(order - 1), vy(order - 1), vz(order - 1);
 
     /* coef of vel */
-    for (int i = 0; i < order - 1; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(order - 1); ++i) {
       vx(i) = double(i + 1) * cxs[idx][order - 2 - i];
       vy(i) = double(i + 1) * cys[idx][order - 2 - i];
       vz(i) = double(i + 1) * czs[idx][order - 2 - i];
     }
     double ts = t;
     Eigen::VectorXd tv(order - 1);
-    for (int i = 0; i < order - 1; ++i)
+    for (size_t i = 0; i < static_cast<size_t>(order - 1); ++i)
       tv(i) = pow(ts, i);
 
     Eigen::Vector3d vel;
@@ -130,14 +130,14 @@ public:
     Eigen::VectorXd ax(order - 2), ay(order - 2), az(order - 2);
 
     /* coef of vel */
-    for (int i = 0; i < order - 2; ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(order - 2); ++i) {
       ax(i) = double((i + 2) * (i + 1)) * cxs[idx][order - 3 - i];
       ay(i) = double((i + 2) * (i + 1)) * cys[idx][order - 3 - i];
       az(i) = double((i + 2) * (i + 1)) * czs[idx][order - 3 - i];
     }
     double ts = t;
     Eigen::VectorXd tv(order - 2);
-    for (int i = 0; i < order - 2; ++i)
+    for (size_t i = 0; i < static_cast<size_t>(order - 2); ++i)
       tv(i) = pow(ts, i);
 
     Eigen::Vector3d acc;
@@ -165,7 +165,7 @@ public:
     length = 0.0;
 
     Eigen::Vector3d p_l = traj_vec3d[0], p_n;
-    for (int i = 1; i < traj_vec3d.size(); ++i) {
+    for (size_t i = 1; i < traj_vec3d.size(); ++i) {
       p_n = traj_vec3d[i];
       length += (p_n - p_l).norm();
       p_l = p_n;
@@ -182,7 +182,7 @@ public:
     double cost = 0.0;
     int order = cxs[0].size();
 
-    for (int s = 0; s < times.size(); ++s) {
+    for (size_t s = 0; s < times.size(); ++s) {
       Eigen::Vector3d um;
       um(0) = 2 * cxs[s][order - 3], um(1) = 2 * cys[s][order - 3], um(2) = 2 * czs[s][order - 3];
       cost += um.squaredNorm() * times[s];
@@ -195,11 +195,11 @@ public:
     double jerk = 0.0;
 
     /* evaluate jerk */
-    for (int s = 0; s < times.size(); ++s) {
+    for (size_t s = 0; s < times.size(); ++s) {
       Eigen::VectorXd cxv(cxs[s].size()), cyv(cys[s].size()), czv(czs[s].size());
       /* convert coefficient */
       int order = cxs[s].size();
-      for (int j = 0; j < order; ++j) {
+      for (size_t j = 0; j < static_cast<size_t>(order); ++j) {
         cxv(j) = cxs[s][order - 1 - j], cyv(j) = cys[s][order - 1 - j], czv(j) = czs[s][order - 1 - j];
       }
       double ts = times[s];
@@ -224,12 +224,12 @@ public:
   void getMeanAndMaxVel(double& mean_v, double& max_v) {
     int num = 0;
     mean_v = 0.0, max_v = -1.0;
-    for (int s = 0; s < times.size(); ++s) {
+    for (size_t s = 0; s < times.size(); ++s) {
       int order = cxs[s].size();
       Eigen::VectorXd vx(order - 1), vy(order - 1), vz(order - 1);
 
       /* coef of vel */
-      for (int i = 0; i < order - 1; ++i) {
+      for (size_t i = 0; i < static_cast<size_t>(order - 1); ++i) {
         vx(i) = double(i + 1) * cxs[s][order - 2 - i];
         vy(i) = double(i + 1) * cys[s][order - 2 - i];
         vz(i) = double(i + 1) * czs[s][order - 2 - i];
@@ -239,7 +239,7 @@ public:
       double eval_t = 0.0;
       while (eval_t < ts) {
         Eigen::VectorXd tv(order - 1);
-        for (int i = 0; i < order - 1; ++i)
+        for (size_t i = 0; i < static_cast<size_t>(order - 1); ++i)
           tv(i) = pow(ts, i);
         Eigen::Vector3d vel;
         vel(0) = tv.dot(vx), vel(1) = tv.dot(vy), vel(2) = tv.dot(vz);
@@ -258,12 +258,12 @@ public:
   void getMeanAndMaxAcc(double& mean_a, double& max_a) {
     int num = 0;
     mean_a = 0.0, max_a = -1.0;
-    for (int s = 0; s < times.size(); ++s) {
+    for (size_t s = 0; s < times.size(); ++s) {
       int order = cxs[s].size();
       Eigen::VectorXd ax(order - 2), ay(order - 2), az(order - 2);
 
       /* coef of acc */
-      for (int i = 0; i < order - 2; ++i) {
+      for (size_t i = 0; i < static_cast<size_t>(order - 2); ++i) {
         ax(i) = double((i + 2) * (i + 1)) * cxs[s][order - 3 - i];
         ay(i) = double((i + 2) * (i + 1)) * cys[s][order - 3 - i];
         az(i) = double((i + 2) * (i + 1)) * czs[s][order - 3 - i];
@@ -273,7 +273,7 @@ public:
       double eval_t = 0.0;
       while (eval_t < ts) {
         Eigen::VectorXd tv(order - 2);
-        for (int i = 0; i < order - 2; ++i)
+        for (size_t i = 0; i < static_cast<size_t>(order - 2); ++i)
           tv(i) = pow(ts, i);
         Eigen::Vector3d acc;
         acc(0) = tv.dot(ax), acc(1) = tv.dot(ay), acc(2) = tv.dot(az);
@@ -301,3 +301,4 @@ PolynomialTraj fastLine4deg(Eigen::Vector3d start, Eigen::Vector3d end, double m
 PolynomialTraj fastLine3deg(Eigen::Vector3d start, Eigen::Vector3d end, double max_vel, double max_acc);
 
 #endif
+
